+++ changelog.md
@@ -20,23 +20,50 @@ Format: *Added, Changed, Fixed, Removed, Security*.
 - 
 
 ---
-## [0.2.12] - 2025-09-14
+
+## [0.2.17] - 2025-09-30
+### Added
+- Lead email template
+
+### Changed
+- price.ts: restructured "Price" type and things relate to it
+
+### Fixed
+- Any type causing build time error
+
+---
+
+## [0.2.16] - 2025-09-30
+### Added
+- Unique quote id library
+
+---
+
+## [0.2.15] - 2025-09-30
+### Added
+- Lead/Quote related database
+- Price calculation Library
+
+---
+
+## [0.2.12] - 2025-09-24
 ### Added
 - Admin page to add/rmv data from database
 
 ---
 
----
-## [0.2.1] - 2025-09-14
+## [0.2.1] - 2025-09-24
 ### Fixed
 - Backend image not loading due to case sensativity
 
 ---
+
 ## [0.2.0] - 2025-09-23
 ### Added
 - supabase backend MVP
 
 ---
+
 ## [0.1.2] - 2025-09-19
 ### Added
 - Add CI/CD pipline via Vercel